+++ lib/router.js
@@ -3097,6 +3097,7 @@ router.get('/soomal/topics/:category/:language?', require('./routes/soomal/topic
 // NASA
 router.get('/nasa/apod', require('./routes/nasa/apod'));
 router.get('/nasa/apod-ncku', require('./routes/nasa/apod-ncku'));
+router.get('/nasa/apod-cn', require('./routes/nasa/apod-cn'));
 
 // 爱Q生活网
 router.get('/iqshw/latest', require('./routes/iqshw/latest'));
@@ -3115,8 +3116,5 @@ router.get('/fjnews/fz/30', require('./routes/fjnews/fznews'));
 
 // 九江新闻jjnews
 router.get('/fjnews/jjnews', require('./routes/fjnews/jjnews'));
-
-// NASA中文
-router.get('/nasachina/apotd', require('./routes/nasachina/apotd'));
 
 module.exports = router;
